library now exports SwzWriter, SwzReader and get_swz_file_name - added example rust files
